+++ changelog.md
@@ -1,6 +1,6 @@
 ### Revision History
 * 4.0.18-SNAPSHOT
-  * --
+  * Enhancement: Can add a reference to an existing axis so as to not lose cell data.
 * 4.0.17
   * Added getters for `NCubeRuntimeClient` and `NCubeMutableClient` on `NCubeGroovyExpression`
   * Bug fix: updated property defaults in Spring beans configuration